+++ deno.json
@@ -1,8 +1,9 @@
 {
   "tasks": {
-    "create:all": "deno run -A ./src/tools/create-all.ts",
-    "create:sebu": "deno run -A ./src/tools/create-sebunagata.ts",
-    "update": "deno run --allow-env --allow-net --allow-read --allow-write ./src/main.ts && deno fmt ./docs/",
-    "gen": "deno run -A ./src/tools/generate.ts"
+    "create:json": "deno run -A ./src/tools/json/create-all.ts & deno run -A ./src/tools/json/create-sebunagata.ts",
+    "create:podcast": "deno run -A ./src/tools/podcast/create-podcast.ts",
+    "update": "deno run --allow-env --allow-net --allow-read --allow-write ./src/main.ts && deno task fmt",
+    "gen": "deno run -A ./src/tools/json/generate.ts",
+    "fmt": "deno fmt ./docs/"
   }
 }
