+++ src/libs/json/add.ts
@@ -1,22 +1,22 @@
-import type { Episode } from "../types/episode.ts";
+import type { Episode } from "../../types/episode.ts";
 
 import { fetchEpisodeInfo } from "./fetch.ts";
 
 /**
  * 新しいエピソードを追加
  * @param url 記事URL
- * @returns 成功したか
+ * @returns 追加先のラジオID
  */
-export async function addEpisode(url: string): Promise<boolean> {
+export async function addEpisode(url: string): Promise<string | undefined> {
   // ラジオ以外の記事ならスキップ
   if (!/radio|rensai/.test(url)) {
     console.log(`[SKIP] ラジオの記事ではありません (${url})`);
-    return false;
+    return undefined;
   }
 
   // 詳細を取得
   const episodeInfo = await fetchEpisodeInfo(url);
-  if (!episodeInfo) return false;
+  if (!episodeInfo) return undefined;
 
   const { id, episodeTitle, episodeNumber, desc, pubDate, source } =
     episodeInfo;
@@ -46,7 +46,7 @@ export async function addEpisode(url: string): Promise<boolean> {
 
   if (isDuplicate) {
     console.log(`[SKIP] 重複したエピソードです (${newEpisode.title})`);
-    return true;
+    return undefined;
   }
 
   // 追加して保存
@@ -55,6 +55,6 @@ export async function addEpisode(url: string): Promise<boolean> {
   const results = episodes.sort((a, b) => a.number - b.number);
   Deno.writeTextFileSync(episodeJsonPath, JSON.stringify(results, null, "\t"));
 
-  console.info("[ADDED]");
-  return true;
+  console.info(`[ADDED] ${id}.json`);
+  return id;
 }
